requirements.txt
pip
Flask
cassandra-driver
pymysql
requests_cache
